+++ quickwit/quickwit-indexing/src/actors/merge_scheduler_service.rs
@@ -228,10 +228,10 @@ impl MergeSchedulerService {
             self.pending_merge_bytes -= merge_task.merge_operation.total_num_bytes();
             crate::metrics::INDEXER_METRICS
                 .pending_merge_operations
-                .set(self.pending_merge_queue.len() as i64);
+                .set(self.pending_merge_queue.len() as f64);
             crate::metrics::INDEXER_METRICS
                 .pending_merge_bytes
-                .set(self.pending_merge_bytes as i64);
+                .set(self.pending_merge_bytes as f64);
             match split_downloader_mailbox.try_send_message(merge_task) {
                 Ok(_) => {}
                 Err(quickwit_actors::TrySendError::Full(_)) => {
@@ -297,7 +297,7 @@ impl MergeSchedulerService {
             self.merge_concurrency as i64 - self.merge_semaphore.available_permits() as i64;
         crate::metrics::INDEXER_METRICS
             .ongoing_merge_operations
-            .set(num_merges);
+            .set(num_merges as f64);
     }
 }
 
@@ -383,10 +383,10 @@ impl Handler<ScheduleMerge> for MergeSchedulerService {
         self.pending_merge_queue.push(scheduled_merge);
         crate::metrics::INDEXER_METRICS
             .pending_merge_operations
-            .set(self.pending_merge_queue.len() as i64);
+            .set(self.pending_merge_queue.len() as f64);
         crate::metrics::INDEXER_METRICS
             .pending_merge_bytes
-            .set(self.pending_merge_bytes as i64);
+            .set(self.pending_merge_bytes as f64);
         self.schedule_pending_merges(ctx);
         Ok(())
     }
